* commit '421d8ba1dd73a67bb4b307f09f352106bd48fc5e': (60 commits)
  CommonStructures.java edited online with Bitbucket
  dhl.wsdl edited online with Bitbucket
  dhl.wsdl edited online with Bitbucket
  TestClient.java edited online with Bitbucket
  getSendStatus_v2.xml edited online with Bitbucket
  runSystemCheck_v1.xml edited online with Bitbucket
  deleteOldDocuments_v1.xml edited online with Bitbucket
  Include dvkcore in client, replace nonexistent XRoadProtocolHeader with XRoadHeader
  objectType in testcases
  objectType to tests
  Changed artifactory url in pom.xml
  dhx-adapter 1.0.1
  special subsystems
  special subsystems
  revert
  special subsystem
  revert
  package with default sender
  DHX duplicate fixed
  fix
  ...

+++ core/src/main/java/dvk/core/CommonMethods.java
@@ -312,6 +312,25 @@ public class CommonMethods {
             throw new RuntimeException(e);
         }
     }
+    
+    /**
+     * Recursively renames the namespace of a node.
+     * 
+     * @param node the starting node.
+     * @param namespace the new namespace. Supplying <tt>null</tt> removes the namespace.
+     */
+    public static void removeNamespaceRecursive(Node node) {
+     // Document document = node.getOwnerDocument();
+      if (node.getNodeType() == Node.ELEMENT_NODE) {
+        node.setPrefix(null);
+       // document.renameNode(node, namespace, node.getNodeName());
+      }
+      NodeList list = node.getChildNodes();
+      for (int i = 0; i < list.getLength(); ++i) {
+        removeNamespaceRecursive(list.item(i));
+      }
+    }
+
 
     public static byte[] xmlElementToBinary(Element xmlElement) {
         try {
@@ -326,6 +345,7 @@ public class CommonMethods {
         boolean result = true;
         FileOutputStream out = null;
         OutputStreamWriter ow = null;
+        xmlElement.setPrefix(null);
         try {
             out = new FileOutputStream(filePath, false);
             ow = new OutputStreamWriter(out, "UTF-8");
@@ -1035,8 +1055,12 @@ public class CommonMethods {
 
         return result;
     }
-
+    
     public static void joinSplitXML(String xmlFileName, BufferedWriter mainOutWriter) {
+    	joinSplitXML(xmlFileName, mainOutWriter, true);
+    }
+
+    public static void joinSplitXML(String xmlFileName, BufferedWriter mainOutWriter, Boolean deleteOriginalFile) {
         FileInputStream mainInStream = null;
         InputStreamReader mainInReader = null;
         BufferedReader mainReader = null;
@@ -1099,9 +1123,10 @@ public class CommonMethods {
             safeCloseReader(mainReader);
             safeCloseReader(mainInReader);
             safeCloseStream(mainInStream);
-
-            // Kustutame ajutise faili
-            (new File(xmlFileName)).delete();
+            if ( deleteOriginalFile) {
+	            // Kustutame ajutise faili
+	            (new File(xmlFileName)).delete();
+            }
         } catch (Exception ex) {
         	logger.error(ex.getMessage(), ex);
         } finally {
